Fixed mapping issue, removed mapstruct

+++ java-crawler/src/main/java/com/hronosf/crawler/services/impl/ElasticSearchWrapperServiceImpl.java
@@ -25,7 +25,7 @@ public class ElasticSearchWrapperServiceImpl implements ElasticSearchWrapperServ
         String ids = parsedPosts.stream().map(post -> post.getId().toString()).collect(Collectors.joining(","));
 
         // query ES to get existing post by id set:
-        Map<Long, WallPost> existedPosts = repository.findAllById(ids, PageRequest.of(0, 100))
+        Map<Integer, WallPost> existedPosts = repository.findAllById(ids, PageRequest.of(0, 100))
                 .getContent().stream()
                 .collect(Collectors.toMap(WallPost::getId, Function.identity()));
